+++ src/main/java/co/wanted/board/global/model/ModifiedEntity.java
@@ -1,9 +1,9 @@
-package co.wanted.board.config.model;
+package co.wanted.board.global.model;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.EntityListeners;
@@ -14,8 +14,8 @@ import java.time.Instant;
 @MappedSuperclass
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EntityListeners(AuditingEntityListener.class)
-public abstract class CreatedEntity {
+public abstract class ModifiedEntity extends CreatedEntity {
 
-    @CreatedDate
-    private Instant createdAt;
+    @LastModifiedDate
+    private Instant updatedAt;
 }
